bot/requirements.txt
Flask==2.2.2
google-cloud-secret-manager==2.14.0
gunicorn==20.1.0
requests==2.28.1
requests-oauthlib==1.3.1
